Support replication when newest remote snapshot no longer exists locally.
This is a generalization of the algorithm.  Verbose output now shows
local and remote snapshots with a `-->' indicator that shows the sync
start point, and a `DEL' indicator that shows which remote snapshots
must be deleted to enable continued replication.  This can happend when
snapshots (e.g. hourlies or dailies, etc.) drop off but older snapshots
still exist.